+++ cvat-core/src/api.ts
@@ -155,6 +155,12 @@ function build() {
                 return result;
             },
         },
+        frames: {
+            async getMeta(type, id) {
+                const result = await PluginRegistry.apiWrapper(cvat.frames.getMeta, type, id);
+                return result;
+            },
+        },
         users: {
             async get(filter = {}) {
                 const result = await PluginRegistry.apiWrapper(cvat.users.get, filter);
@@ -272,6 +278,24 @@ function build() {
                 return result;
             },
         },
+        analytics: {
+            quality: {
+                async reports(filter: any) {
+                    const result = await PluginRegistry.apiWrapper(cvat.analytics.quality.reports, filter);
+                    return result;
+                },
+                async conflicts(filter: any) {
+                    const result = await PluginRegistry.apiWrapper(cvat.analytics.quality.conflicts, filter);
+                    return result;
+                },
+                settings: {
+                    async get(taskID: number) {
+                        const result = await PluginRegistry.apiWrapper(cvat.analytics.quality.settings.get, taskID);
+                        return result;
+                    },
+                },
+            },
+        },
         classes: {
             User,
             Project: implementProject(Project),
@@ -304,6 +328,8 @@ function build() {
     cvat.enums = Object.freeze(cvat.enums);
     cvat.cloudStorages = Object.freeze(cvat.cloudStorages);
     cvat.organizations = Object.freeze(cvat.organizations);
+    cvat.analytics = Object.freeze(cvat.analytics);
+    cvat.frames = Object.freeze(cvat.frames);
 
     const implemented = Object.freeze(implementAPI(cvat));
     return implemented;